A bunch of stuff
Messed around with lighting and created a 'TestLevel' due to issues with lighting. Also found out how to import objects correctly! use import as dynamic when the dialogue appears. Also played about with the terrain tools on the TestLevel. May be worth creating a brand new level and get to sculpting the terrain. The MainLevel level can just be a prototype for now (but there's many issues with it and rendering the lighting takes an ETERNITY for unknown reasons)